requirements.txt
alembic==1.10.4
autopep8==2.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
python-dotenv==1.0.0
requests==2.29.0
soupsieve==2.4.1
SQLAlchemy==2.0.10
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.0
youtube-transcript-api==0.6.0
